+++ tests/test_ml_tools.py
@@ -14,22 +14,24 @@ def prepare_jets_array():
             "pt": ak.from_numpy(np.random.random(size=NJETS)),
             "eta": ak.from_numpy(np.random.random(size=NJETS)),
             "phi": ak.from_numpy(np.random.random(size=NJETS)),
-            "ntrk": ak.from_numpy(np.random.randint(1, 50, size=NJETS)),
-        }
+            "ncands": ak.from_numpy(np.random.randint(1, 50, size=NJETS)),
+        },
+        with_name="LorentzVector",
     )
-    tracks = ak.zip(
+    pfcands = ak.zip(
         {
             "pt": ak.from_regular(np.random.random(size=(NJETS, NFEAT))),
             "eta": ak.from_regular(np.random.random(size=(NJETS, NFEAT))),
             "phi": ak.from_regular(np.random.random(size=(NJETS, NFEAT))),
-            "ip2d": ak.from_regular(np.random.random(size=(NJETS, NFEAT))),
-            "ipz": ak.from_regular(np.random.random(size=(NJETS, NFEAT))),
-        }
+            "feat1": ak.from_regular(np.random.random(size=(NJETS, NFEAT))),
+            "feat2": ak.from_regular(np.random.random(size=(NJETS, NFEAT))),
+        },
+        with_name="LorentzVector",
     )
 
-    idx = ak.local_index(tracks.pt, axis=-1)
-    tracks = tracks[idx < jets.ntrk]
-    jets["tracks"] = tracks[:]
+    idx = ak.local_index(pfcands.pt, axis=-1)
+    pfcands = pfcands[idx < jets.ncands]
+    jets["pfcands"] = pfcands[:]
 
     ak_jets = jets[:]
     ak.to_parquet(jets, "ml_tools.parquet")
@@ -37,56 +39,59 @@ def prepare_jets_array():
     return ak_jets, dak_jets
 
 
-def triton_testing():
+def common_awkward_to_numpy(jets):
+    def my_pad(arr):
+        return ak.fill_none(ak.pad_none(arr, 100, axis=1, clip=True), 0.0)
+
+    fmap = {
+        "points__0": {
+            "deta": my_pad(jets.eta - jets.pfcands.eta),
+            "dphi": my_pad(jets.phi - jets.pfcands.phi),
+        },
+        "features__1": {
+            "dr": my_pad(
+                np.sqrt(
+                    (jets.eta - jets.pfcands.eta) ** 2
+                    + (jets.phi - jets.pfcands.phi) ** 2
+                )
+            ),
+            "lpt": my_pad(np.log(jets.pfcands.pt)),
+            "lptf": my_pad(np.log(jets.pfcands.pt / ak.sum(jets.pfcands.pt, axis=-1))),
+            "f1": my_pad(np.log(jets.pfcands.feat1 + 1)),
+            "f2": my_pad(np.log(jets.pfcands.feat2 + 1)),
+        },
+        "mask__2": {
+            "mask": my_pad(ak.ones_like(jets.pfcands.pt)),
+        },
+    }
+
+    return {
+        k: ak.concatenate(
+            [x[:, np.newaxis, :] for x in fmap[k].values()], axis=1
+        ).to_numpy()
+        for k in fmap.keys()
+    }
+
+
+def test_triton():
     # Defining custom wrapper function with awkward padding requirements.
     class triton_wrapper_test(coffea.ml_tools.triton_wrapper):
-        def awkward_to_numpy(self, jets):
-            def my_pad(arr):
-                return ak.fill_none(ak.pad_none(arr, 100, axis=1, clip=True), 0.0)
-
-            fmap = {
-                "points__0": {
-                    "deta": my_pad(jets.eta - jets.tracks.eta),
-                    "dphi": my_pad(jets.phi - jets.tracks.phi),
-                },
-                "features__1": {
-                    "dr": my_pad(
-                        np.sqrt(
-                            (jets.eta - jets.tracks.eta) ** 2
-                            + (jets.phi - jets.tracks.phi) ** 2
-                        )
-                    ),
-                    "lpt": my_pad(np.log(jets.tracks.pt)),
-                    "lptf": my_pad(
-                        np.log(jets.tracks.pt / ak.sum(jets.tracks.pt, axis=-1))
-                    ),
-                    "ip2d": my_pad(
-                        np.sign(jets.tracks.ip2d) * np.log(jets.tracks.ip2d + 1)
-                    ),
-                    "ipz": my_pad(
-                        np.sign(jets.tracks.ipz) * np.log(jets.tracks.ipz + 1)
-                    ),
-                },
-                "mask__2": {
-                    "mask": my_pad(ak.ones_like(jets.tracks.pt)),
-                },
+        def awkward_to_numpy(self, output_list, jets):
+            return [], {
+                "output_list": output_list,
+                "input_dict": common_awkward_to_numpy(jets),
             }
 
-            return {
-                k: ak.concatenate(
-                    [x[:, np.newaxis, :] for x in fmap[k].values()], axis=1
-                ).to_numpy()
-                for k in fmap.keys()
-            }
-
-        def dask_touch(self, jets):
-            jets.eta.layout._touch_data(recursive=False)
-            jets.phi.layout._touch_data(recursive=False)
-            jets.tracks.pt.layout._touch_data(recursive=False)
-            jets.tracks.eta.layout._touch_data(recursive=False)
-            jets.tracks.ip2d.layout._touch_data(recursive=False)
-            jets.tracks.ipz.layout._touch_data(recursive=False)
-            pass
+        def dask_columns(self, output_list, jets):
+            return [
+                jets.eta,
+                jets.phi,
+                jets.pfcands.pt,
+                jets.pfcands.phi,
+                jets.pfcands.eta,
+                jets.pfcands.feat1,
+                jets.pfcands.feat2,
+            ]
 
     # Running the evaluation in lazy and non-lazy forms
     tw = triton_wrapper_test(
@@ -96,7 +101,7 @@ def triton_testing():
     ak_jets, dak_jets = prepare_jets_array()
 
     # Numpy arrays testing
-    np_res = tw._numpy_call(["softmax__0"], tw.awkward_to_numpy(ak_jets), validate=True)
+    np_res = tw(["softmax__0"], common_awkward_to_numpy(ak_jets))
     print({k: v.shape for k, v in np_res.items()})
 
     # Vanilla awkward arrays
@@ -113,7 +118,3 @@ def triton_testing():
     for k in ak_res.keys():
         assert ak.all(ak_res[k] == dak_res[k].compute())
     print(dak.necessary_columns(dak_res))
-
-
-if __name__ == "__main__":
-    triton_testing()
